+++ Backend/service/material.js
@@ -14,10 +14,28 @@ async function createNewMaterial(newMaterial){
 }
 
 async function updateMaterial(updateMaterial){
+    const isExist = await material.getMaterialById(updateMaterial.id);
+
+    if(isExist.data === null)
+        return {
+            operation:  false,
+            status:     404,
+            message:    `Material with id ${updateMaterial.id} was not found`
+        }
+
     return await material.updateMaterialById(updateMaterial);
 }
 
 async function deleteMaterial(id){
+    const isExist = await material.getMaterialById(id);
+
+    if(isExist.data === null)
+        return {
+            operation:  false,
+            status:     404,
+            message:    `Material with id ${id} was not found`
+        }
+
     return await material.deleteMaterialById(id);
 }
